update: HOME＞メモ別学習効率のレスポンシブデザイン対応
